% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap_wrappers.R
\name{as_sap}
\alias{as_sap}
\title{Convert a LYS object to an ASAP Sap object}
\usage{
as_sap(x)
}
\arguments{
\item{x}{A \code{lys} object.}
}
\value{
An ASAP \code{Sap} object.
}
\description{
Adapts LYS recording metadata for ASAP functions. In particular, LYS
relative directories become ASAP's \code{day_post_hatch} paths, allowing
ASAP to find WAV files below the LYS base directory.
}
\examples{
\dontrun{
sap <- as_sap(lys)
visualize_song(sap, n_samples = 4, random = TRUE)
}

}
