% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{assign_recording_sessions}
\alias{assign_recording_sessions}
\title{Assign recording sessions to a LYS metadata table}
\usage{
assign_recording_sessions(metadata, session_gap_hours = 1)
}
\arguments{
\item{metadata}{Data frame as returned by \code{create_lys_metadata()}.}

\item{session_gap_hours}{Numeric. Minimum gap in hours between sessions.
Default \code{1}.}
}
\value{
The metadata data frame with session columns added.
}
\description{
Assign recording sessions to a LYS metadata table
}
