% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{parse_sap_filename}
\alias{parse_sap_filename}
\title{Parse a SAP-format filename into metadata fields}
\usage{
parse_sap_filename(filename, tz = "UTC")
}
\arguments{
\item{filename}{Character. Path or filename to parse.}

\item{tz}{Character. Timezone string. Default \code{"UTC"}.}
}
\value{
A named list of metadata fields.
}
\description{
Parse a SAP-format filename into metadata fields
}
