% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.lys}
\alias{print.lys}
\title{Print method for LYS object}
\usage{
\method{print}{lys}(x, ...)
}
\arguments{
\item{x}{A \code{lys} object}

\item{...}{Unused}
}
\value{
The \code{lys} object, invisibly.
}
\description{
Display a concise overview of a \code{lys} object.
}
\keyword{internal}
