% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{summary.lys}
\alias{summary.lys}
\title{Summary method for LYS object}
\usage{
\method{summary}{lys}(object, ...)
}
\arguments{
\item{object}{A \code{lys} object}

\item{...}{Unused}
}
\value{
The \code{lys} object, invisibly.
}
\description{
Summarize the contents, session structure, and detections in a \code{lys} object.
}
